Eliminated a bunch of compiler warnings.

+++ epoch1d/src/housekeeping/balance.F90
@@ -22,9 +22,9 @@ CONTAINS
     ! So cheat
 
     LOGICAL, INTENT(IN) :: over_ride
-    INTEGER(KIND=8), DIMENSION(:), ALLOCATABLE :: load_x
+    INTEGER(i8), DIMENSION(:), ALLOCATABLE :: load_x
     REAL(num) :: balance_frac, npart_av
-    INTEGER(KIND=8) :: npart_local, sum_npart, max_npart
+    INTEGER(i8) :: npart_local, sum_npart, max_npart
     INTEGER :: iproc
     INTEGER, DIMENSION(c_ndims,2) :: domain
 #ifdef PARTICLE_DEBUG
@@ -403,8 +403,8 @@ CONTAINS
 
     ! Calculate total load across the X direction
 
-    INTEGER(KIND=8), DIMENSION(:), INTENT(OUT) :: load
-    INTEGER(KIND=8), DIMENSION(:), ALLOCATABLE :: temp
+    INTEGER(i8), DIMENSION(:), INTENT(OUT) :: load
+    INTEGER(i8), DIMENSION(:), ALLOCATABLE :: temp
     TYPE(particle), POINTER :: current
     INTEGER :: cell, ispecies, sz
 
@@ -445,11 +445,11 @@ CONTAINS
     ! This subroutine calculates the places in a given load profile to split
     ! The domain to give the most even subdivision possible
 
-    INTEGER(KIND=8), INTENT(IN), DIMENSION(:) :: load
+    INTEGER(i8), INTENT(IN), DIMENSION(:) :: load
     INTEGER, INTENT(IN) :: nproc
     INTEGER, DIMENSION(:), INTENT(OUT) :: mins, maxs
     INTEGER :: sz, idim, proc, old
-    INTEGER(KIND=8) :: total, total_old, load_per_proc_ideal
+    INTEGER(i8) :: total, total_old, load_per_proc_ideal
 
     sz = SIZE(load)
     maxs = sz
@@ -547,7 +547,7 @@ CONTAINS
     TYPE(particle_list), DIMENSION(:), ALLOCATABLE :: pointers_recv
     TYPE(particle), POINTER :: current, next
     INTEGER :: part_proc, iproc, ispecies, ierr
-    INTEGER(KIND=8), DIMENSION(:), ALLOCATABLE :: sendcounts, recvcounts
+    INTEGER(i8), DIMENSION(:), ALLOCATABLE :: sendcounts, recvcounts
 
     ALLOCATE(pointers_send(0:nproc-1), pointers_recv(0:nproc-1))
     ALLOCATE(sendcounts(0:nproc-1), recvcounts(0:nproc-1))
