+++ epoch3d/src/io/calc_df.F90
@@ -882,7 +882,7 @@ CONTAINS
     REAL(num) :: root, fac, third, gamma, cf2
     REAL(num) :: delta_x, delta_y, delta_z
     INTEGER :: ispecies, ix, iy, iz, dcellx, dcelly, dcellz
-    INTEGER(KIND=8) :: ipart
+    INTEGER(i8) :: ipart
     TYPE(particle), POINTER :: current, next
     INTEGER :: spec_start, spec_end
     LOGICAL :: spec_sum
